+++ gazebo/common/Color.cc
@@ -17,7 +17,6 @@
 #include <math.h>
 #include <algorithm>
 
-#include "gazebo/math/Helpers.hh"
 #include "gazebo/common/Console.hh"
 #include "gazebo/common/Color.hh"