+++ gazebo/physics/dart/DARTUniversalJoint.cc
@@ -28,19 +28,21 @@ using namespace physics;
 DARTUniversalJoint::DARTUniversalJoint(BasePtr _parent)
   : UniversalJoint<DARTJoint>(_parent)
 {
-  this->dataPtr->dtJoint = new dart::dynamics::UniversalJoint();
 }
 
 //////////////////////////////////////////////////
 DARTUniversalJoint::~DARTUniversalJoint()
 {
-  // We don't need to delete dtJoint because the world will delete it
 }
 
 //////////////////////////////////////////////////
 void DARTUniversalJoint::Load(sdf::ElementPtr _sdf)
 {
   UniversalJoint<DARTJoint>::Load(_sdf);
+
+  this->dataPtr->dtProperties.reset(
+        new dart::dynamics::UniversalJoint::Properties(
+          *this->dataPtr->dtProperties.get()));
 }
 
 //////////////////////////////////////////////////
@@ -50,8 +52,14 @@ void DARTUniversalJoint::Init()
 }
 
 //////////////////////////////////////////////////
-math::Vector3 DARTUniversalJoint::GetAnchor(unsigned int /*index*/) const
+math::Vector3 DARTUniversalJoint::GetAnchor(unsigned int _index) const
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    return this->dataPtr->GetCached<math::Vector3>(
+          "Anchor" + std::to_string(_index));
+  }
+
   Eigen::Isometry3d T = this->dataPtr->dtChildBodyNode->getTransform() *
                         this->dataPtr->dtJoint->getTransformFromChildBodyNode();
   Eigen::Vector3d worldOrigin = T.translation();
@@ -62,6 +70,12 @@ math::Vector3 DARTUniversalJoint::GetAnchor(unsigned int /*index*/) const
 //////////////////////////////////////////////////
 math::Vector3 DARTUniversalJoint::GetGlobalAxis(unsigned int _index) const
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    return this->dataPtr->GetCached<math::Vector3>(
+          "Axis" + std::to_string(_index));
+  }
+
   Eigen::Vector3d globalAxis = Eigen::Vector3d::UnitX();
 
   if (_index == 0)
@@ -101,6 +115,14 @@ math::Vector3 DARTUniversalJoint::GetGlobalAxis(unsigned int _index) const
 void DARTUniversalJoint::SetAxis(unsigned int _index,
     const math::Vector3 &_axis)
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    this->dataPtr->Cache(
+          "Axis" + std::to_string(_index),
+          boost::bind(&DARTUniversalJoint::SetAxis, this, _index, _axis));
+    return;
+  }
+
   Eigen::Vector3d dtAxis = DARTTypes::ConvVec3(
       this->GetAxisFrameOffset(_index).RotateVector(_axis));
   Eigen::Isometry3d dtTransfJointLeftToParentLink
@@ -130,6 +152,12 @@ void DARTUniversalJoint::SetAxis(unsigned int _index,
 //////////////////////////////////////////////////
 math::Angle DARTUniversalJoint::GetAngleImpl(unsigned int _index) const
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    return this->dataPtr->GetCached<math::Angle>(
+          "Angle" + std::to_string(_index));
+  }
+
   math::Angle result;
 
   if (_index == 0)
@@ -153,6 +181,12 @@ math::Angle DARTUniversalJoint::GetAngleImpl(unsigned int _index) const
 //////////////////////////////////////////////////
 double DARTUniversalJoint::GetVelocity(unsigned int _index) const
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    return this->dataPtr->GetCached<double>(
+          "Velocity" + std::to_string(_index));
+  }
+
   double result = 0.0;
 
   if (_index == 0)
@@ -168,12 +202,17 @@ double DARTUniversalJoint::GetVelocity(unsigned int _index) const
 //////////////////////////////////////////////////
 void DARTUniversalJoint::SetVelocity(unsigned int _index, double _vel)
 {
-  if (_index < this->GetAngleCount())
+  if (!this->dataPtr->IsInitialized())
   {
-    this->dataPtr->dtJoint->setVelocity(_index, _vel);
-    this->dataPtr->dtJoint->getSkeleton()->computeForwardKinematics(
-          false, true, false);
+    this->dataPtr->Cache(
+          "Velocity" + std::to_string(_index),
+          boost::bind(&DARTUniversalJoint::SetVelocity, this, _index, _vel),
+          _vel);
+    return;
   }
+
+  if (_index < this->GetAngleCount())
+    this->dataPtr->dtJoint->setVelocity(_index, _vel);
   else
     gzerr << "Invalid index[" << _index << "]\n";
 }
@@ -181,6 +220,14 @@ void DARTUniversalJoint::SetVelocity(unsigned int _index, double _vel)
 //////////////////////////////////////////////////
 void DARTUniversalJoint::SetForceImpl(unsigned int _index, double _effort)
 {
+  if (!this->dataPtr->IsInitialized())
+  {
+    this->dataPtr->Cache(
+        "Force" + std::to_string(_index),
+        boost::bind(&DARTUniversalJoint::SetForceImpl, this, _index, _effort));
+    return;
+  }
+
   if (_index == 0)
     this->dataPtr->dtJoint->setForce(0, _effort);
   else if (_index == 1)